Refactor: Update Settings screen UI and logic
