+++ presentation/profile/src/main/java/com/giraffe/presentation/profile/screens/history/HistoryInteractionListener.kt
@@ -4,5 +4,6 @@ interface HistoryInteractionListener {
     fun onDeleteClicked(id: Int, mediaType: String)
     fun onCloseClicked()
     fun onMediaClicked(mediaId: Int, mediaType: String)
+    fun onBackClick()
     fun navigateToExploreScreen()
 }
