+++ presentation/profile/src/main/java/com/giraffe/presentation/profile/screens/settings/SettingsEffect.kt
@@ -1,7 +1,10 @@
 package com.giraffe.presentation.profile.screens.settings
 
 sealed interface SettingsEffect {
+    data object NavigateToEditProfileWebView : SettingsEffect
     data object NavigateToLogin : SettingsEffect
+    data object NavigateToMyCollections : SettingsEffect
+    data object NavigateToHistory : SettingsEffect
+    data object NavigateToRatings : SettingsEffect
     data class ShowError(val error: Throwable) : SettingsEffect
-    data object NavigateToEditProfileWebsite : SettingsEffect
 }
