+++ presentation/profile/src/main/java/com/giraffe/presentation/profile/screens/settings/SettingsInteractionListener.kt
@@ -11,8 +11,10 @@ interface SettingsInteractionListener {
     fun onConfirmLogout()
     fun onGoToWebsiteClick()
     fun onDismissSheet()
-    fun onCancelClick()
     fun onContentPreferencesClick()
     fun onLanguageChange(languageCode: String)
     fun onContentPreferenceChange(preference: ContentPreference)
+    fun onNavigateToMyCollections()
+    fun onNavigateToHistory()
+    fun onNavigateToRatings()
 }
