+++ presentation/profile/src/main/java/com/giraffe/presentation/profile/screens/settings/components/SettingsSection.kt
@@ -1,0 +1,37 @@
+package com.giraffe.presentation.profile.screens.settings.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.unit.dp
+import com.giraffe.designsystem.composable.custom.Text
+import com.giraffe.designsystem.theme.Theme
+
+@Composable
+fun SettingsSection(
+    modifier: Modifier = Modifier,
+    title: String,
+    content: @Composable () -> Unit
+) {
+    Column(
+        verticalArrangement = Arrangement.spacedBy(space = 12.dp),
+        modifier = modifier,
+    ) {
+        Text(
+            text = title,
+            style = Theme.textStyle.title.sm,
+            color = Theme.color.shade.primary,
+        )
+        Column(
+            modifier = Modifier
+                .clip(shape = RoundedCornerShape(Theme.radius.lg))
+                .background(Theme.color.background.card)
+        ) {
+            content()
+        }
+    }
+}
